Added MMD Shapekeys to the basic list
Added the shapekeys used by Miku-Miku-Dance to the basic markers .txt list. Interestingly after looking at the viwi and novabeast it turns out that most avatars use a subset. I might refine this liast in the near future (the full list is MUCH longer, I just took the viwi's as reference for now)