Fixed Shapekey Naming Error
When importing shapekeys from a txt file - which will use newline characters - the newline characters were being imported into the shapekey names. This mean that animations in unity would fail to sync up as the names would not be an exact match.

I also respaced things a little bit in the txt file to make things more legible in timelines as to where different sections were.